added tabbed bar choice
